🔧 refactor(interpreters): modularize interpreter classes
- Move BaseInterpreter and MermaidInterpreter to interpreters.py
- Introduce Interpreters list for managing interpreter instances
- Adjust imports and references in related files to use new module

🗑️ chore(types): extract content types to separate module

- Extract ContentTypes and related classes to type.py
- Update imports and references to use new type module
