+++ playground/telegramify_case.py
@@ -6,7 +6,7 @@ from dotenv import load_dotenv
 from telebot import TeleBot
 
 import telegramify_markdown
-from telegramify_markdown import ContentTypes
+from telegramify_markdown.interpreters import ContentTypes
 from telegramify_markdown.customize import markdown_symbol
 
 tips = """
